create jmock plugin with all extlib (to load hamcrest before org.junit), move all tests to the test plugin and fix tests for Windows OS file separators.
