+++ com.rescripter.editor.test/src/com/rescripter/test/matchers/PathMatcher.java
