+++ com.rescripter.editor.test/src/com/rescripter/test/resources/MockContainerBuilder.java
