+++ com.rescripter.editor/build.properties
@@ -6,10 +6,3 @@ bin.includes = plugin.xml,\
                icons/,\
                lib/rhino-1.7R2.jar,\
                resources/
-jars.extra.classpath = extlib/jmock-junit4-2.5.1.jar,\
-                       extlib/jmock-legacy-2.5.1.jar,\
-                       extlib/jmock-2.5.1.jar,\
-                       extlib/cglib-nodep-2.1_3.jar,\
-                       extlib/objenesis-1.0.jar,\
-                       extlib/hamcrest-core-1.2.1.jar,\
-                       extlib/hamcrest-library-1.2.1.jar
