+++ com.rescripter.editor/src/com/rescripter/script/ScriptRunner.java
@@ -5,6 +5,7 @@ import java.io.InputStreamReader;
 import java.io.LineNumberReader;
 
 import org.eclipse.core.resources.IFile;
+import org.eclipse.core.runtime.Platform;
 import org.eclipse.ui.IWorkbenchWindow;
 import org.mozilla.javascript.Context;
 import org.mozilla.javascript.Scriptable;
@@ -18,43 +19,45 @@ public class ScriptRunner {
     private final Context context;
     private final Scriptable scope;
     private final ScriptLoader scriptLoader;
-	private final DebugMessage debugMessage;
+	private DebugMessage debugMessage;
 
     public ScriptRunner() {
         context = Context.enter();
         scope = context.initStandardObjects();
         putProperty("Load", scriptLoader = new ScriptLoader(this));
+        if (Platform.isRunning()) {
 		putProperty("Debug", debugMessage = new DebugMessage());
+        }
 		putProperty("TestResult", new TestResultPublisher());
     }
-    
+
     public void run(String source, String sourceName, IFile location) {
         scriptLoader.setCurrentLocation(location);
         context.evaluateString(scope, source, sourceName, 1, null);
     }
-    
+
     public void putProperty(String name, Object object) {
         ScriptableObject.putProperty(scope, name, object);
     }
-    
+
     public static ScriptRunner createBasicScriptRunner() {
     	return new ScriptRunner();
     }
-    
+
     public static ScriptRunner createJavaSyntaxScriptRunner(IWorkbenchWindow window) throws IOException {
         Alerter alerter = new Alerter(window);
-        
+
         ScriptRunner runner = new ScriptRunner();
         runner.putProperty("Alert", alerter);
         runner.putProperty("SearchHelper", new SearchHelper());
         runner.putProperty("ChangeText", new ChangeText());
         runner.putProperty("ASTTokenFinder", new ASTTokenFinder());
-        
+
         runner.includeSystem();
-        
+
         return runner;
     }
-    
+
     private void includeSystem() throws IOException {
     	StringBuffer buff = new StringBuffer();
     	LineNumberReader in = new LineNumberReader(new InputStreamReader(getClass().getClassLoader().getResourceAsStream("System.rs")));
