+++ com.rescripter.jmock/.classpath
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/J2SE-1.5"/>
+	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
+	<classpathentry exported="true" kind="lib" path="lib/hamcrest-core-1.2.1.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/hamcrest-library-1.2.1.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jmock-2.5.1.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jmock-junit4-2.5.1.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/jmock-legacy-2.5.1.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/cglib-nodep-2.1_3.jar"/>
+	<classpathentry exported="true" kind="lib" path="lib/objenesis-1.0.jar"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
