+++ com.rescripter.jmock/META-INF/MANIFEST.MF
@@ -1,0 +1,33 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: JMock Dependencies
+Bundle-SymbolicName: com.rescripter.jmock;singleton:=true
+Bundle-Version: 1.0.7.qualifier
+Bundle-RequiredExecutionEnvironment: J2SE-1.5
+Bundle-ClassPath: lib/hamcrest-core-1.2.1.jar,
+ lib/hamcrest-library-1.2.1.jar,
+ lib/jmock-2.5.1.jar,
+ lib/jmock-junit4-2.5.1.jar,
+ lib/jmock-legacy-2.5.1.jar,
+ lib/cglib-nodep-2.1_3.jar,
+ lib/objenesis-1.0.jar
+Export-Package: org.hamcrest,
+ org.hamcrest.beans,
+ org.hamcrest.collection,
+ org.hamcrest.core,
+ org.hamcrest.internal,
+ org.hamcrest.number,
+ org.hamcrest.object,
+ org.hamcrest.text,
+ org.hamcrest.xml,
+ org.jmock,
+ org.jmock.api,
+ org.jmock.integration.junit4,
+ org.jmock.internal,
+ org.jmock.internal.matcher,
+ org.jmock.lib,
+ org.jmock.lib.action,
+ org.jmock.lib.concurrent,
+ org.jmock.lib.concurrent.internal,
+ org.jmock.lib.legacy,
+ org.jmock.syntax
