+++ com.rescripter.jmock/build.properties
@@ -1,0 +1,2 @@
+bin.includes = META-INF/,\
+               lib/
