+++ com.rescripter.jmock/pom.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+    <parent>
+        <groupId>com.rescripter</groupId>
+        <artifactId>Rescripter</artifactId>
+        <version>1.0.7-SNAPSHOT</version>
+    </parent>
+
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>com.rescripter.jmock</artifactId>
+    <packaging>eclipse-plugin</packaging>
+    <name>Rescripter JMock Dependencies</name>
+
+</project>
